// EXERCISES/9.TypeScript/d.React with Types/Exercises-9.14-9.15/src/types.d.ts
interface CoursePartBase {
    name: string;
    exerciseCount: number;
    type: string;
  }
interface CoursePartWithDescription extends CoursePartBase {
    description: string
}

  
  interface CourseNormalPart extends CoursePartWithDescription {
    type: "normal";
  }
  interface CourseProjectPart extends CoursePartBase {
    type: "groupProject";
    groupProjectCount: number;
  }
  
  interface CourseSubmissionPart extends CoursePartWithDescription {
    type: "submission";
    exerciseSubmissionLink: string;
  }
  interface CourseSpecialPart extends CoursePartWithDescription {
    type: "special"
    requirements: ["nodejs", "jest"]
  }
  
export type CoursePart = CourseNormalPart | CourseProjectPart | CourseSubmissionPart | CourseSpecialPart;
  